+++ addons/im_livechat/__manifest__.py
@@ -69,7 +69,8 @@ Help your customers with this chat, and analyse their feedback.
         ],
         'web.qunit_suite_tests': [
             'im_livechat/static/tests/qunit_suite_tests/components/**/*.js',
-            'im_livechat/static/src/legacy/public_livechat.js',
+            'im_livechat/static/src/legacy/public_livechat_constants.js',
+            'im_livechat/static/src/legacy/public_livechat_history_tracking.js',
             'im_livechat/static/src/legacy/models/*',
             'im_livechat/static/src/legacy/widgets/*',
             'im_livechat/static/src/legacy/public_livechat_chatbot.js',
@@ -174,7 +175,8 @@ Help your customers with this chat, and analyse their feedback.
             'bus/static/src/js/crosstab_bus.js',
             'bus/static/src/js/services/bus_service.js',
             'mail/static/src/js/utils.js',
-            'im_livechat/static/src/legacy/public_livechat.js',
+            'im_livechat/static/src/legacy/public_livechat_constants.js',
+            'im_livechat/static/src/legacy/public_livechat_history_tracking.js',
             'im_livechat/static/src/legacy/models/*',
             'im_livechat/static/src/legacy/widgets/*',
             'im_livechat/static/src/legacy/public_livechat_chatbot.js',
