+++ addons/im_livechat/static/src/legacy/website_livechat_message_chatbot.js
@@ -10,7 +10,7 @@ import WebsiteLivechatMessage from 'im_livechat.legacy.im_livechat.model.Website
  */
 WebsiteLivechatMessage.include({
     /**
-     * @param {im_livechat.legacy.im_livechat.im_livechat.LivechatButton} parent
+     * @param {im_livechat.legacy.im_livechat.LivechatButton} parent
      * @param {Object} data
      * @param {Object} options
      * @param {string} options.default_username
