+++ addons/im_livechat/static/src/legacy/widgets/livechat_button.js
@@ -1,8 +1,6 @@
-odoo.define('im_livechat.legacy.im_livechat.im_livechat', function (require) {
+odoo.define('im_livechat.legacy.im_livechat.LivechatButton', function (require) {
 "use strict";
 
-require('bus.BusService');
-var concurrency = require('web.concurrency');
 var config = require('web.config');
 var core = require('web.core');
 var session = require('web.session');
@@ -10,37 +8,14 @@ var time = require('web.time');
 var utils = require('web.utils');
 var Widget = require('web.Widget');
 
+var { LIVECHAT_COOKIE_HISTORY } = require('im_livechat.legacy.im_livechat.Constants');
+var Feedback = require('im_livechat.legacy.im_livechat.Feedback');
 var WebsiteLivechat = require('im_livechat.legacy.im_livechat.model.WebsiteLivechat');
 var WebsiteLivechatMessage = require('im_livechat.legacy.im_livechat.model.WebsiteLivechatMessage');
 var WebsiteLivechatWindow = require('im_livechat.legacy.im_livechat.WebsiteLivechatWindow');
 
 var _t = core._t;
 var QWeb = core.qweb;
-
-// Constants
-var LIVECHAT_COOKIE_HISTORY = 'im_livechat_history';
-var HISTORY_LIMIT = 15;
-
-var RATING_TO_EMOJI = {
-    "5": "😊",
-    "3": "😐",
-    "1": "😞"
-};
-
-// History tracking
-var page = window.location.href.replace(/^.*\/\/[^/]+/, '');
-var pageHistory = utils.get_cookie(LIVECHAT_COOKIE_HISTORY);
-var urlHistory = [];
-if (pageHistory) {
-    urlHistory = JSON.parse(pageHistory) || [];
-}
-if (!_.contains(urlHistory, page)) {
-    urlHistory.push(page);
-    while (urlHistory.length > HISTORY_LIMIT) {
-        urlHistory.shift();
-    }
-    utils.set_cookie(LIVECHAT_COOKIE_HISTORY, JSON.stringify(urlHistory), 60 * 60 * 24); // 1 day cookie
-}
 
 var LivechatButton = Widget.extend({
     className: 'openerp o_livechat_button d-print-none',
@@ -496,151 +471,6 @@ var LivechatButton = Widget.extend({
     },
 });
 
-/*
- * Rating for Livechat
- *
- * This widget displays the 3 rating smileys, and a textarea to add a reason
- * (only for red smiley), and sends the user feedback to the server.
- */
-var Feedback = Widget.extend({
-    template: 'im_livechat.legacy.im_livechat.FeedBack',
-
-    events: {
-        'click .o_livechat_rating_choices img': '_onClickSmiley',
-        'click .o_livechat_no_feedback span': '_onClickNoFeedback',
-        'click .o_rating_submit_button': '_onClickSend',
-        'click .o_email_chat_button': '_onEmailChat',
-        'click .o_livechat_email_error .alert-link': '_onTryAgain',
-    },
-
-    /**
-     * @param {?} parent
-     * @param {im_livechat.legacy.im_livechat.model.WebsiteLivechat} livechat
-     */
-    init: function (parent, livechat) {
-        this._super(parent);
-        this._livechat = livechat;
-        this.server_origin = session.origin;
-        this.rating = undefined;
-        this.dp = new concurrency.DropPrevious();
-    },
-
-    //--------------------------------------------------------------------------
-    // Private
-    //--------------------------------------------------------------------------
-
-    /**
-     * @private
-     * @param {Object} options
-     */
-    _sendFeedback: function (reason) {
-        var self = this;
-        var args = {
-            uuid: this._livechat.getUUID(),
-            rate: this.rating,
-            reason: reason,
-        };
-        this.dp.add(session.rpc('/im_livechat/feedback', args)).then(function (response) {
-            var emoji = RATING_TO_EMOJI[self.rating] || "??";
-            if (!reason) {
-                var content = _.str.sprintf(_t("Rating: %s"), emoji);
-            }
-            else {
-                var content = "Rating reason: \n" + reason;
-            }
-            self.trigger('send_message', { content: content, isFeedback: true });
-        });
-    },
-    /**
-    * @private
-    */
-    _showThanksMessage: function () {
-        this.$('.o_livechat_rating_box').empty().append($('<div />', {
-            text: _t('Thank you for your feedback'),
-            class: 'text-muted'
-        }));
-    },
-
-    //--------------------------------------------------------------------------
-    // Handlers
-    //--------------------------------------------------------------------------
-
-    /**
-     * @private
-     */
-    _onClickNoFeedback: function () {
-        this.trigger('feedback_sent'); // will close the chat
-    },
-    /**
-     * @private
-     */
-    _onClickSend: function () {
-        this.$('.o_livechat_rating_reason').hide();
-        this._showThanksMessage();
-        if (_.isNumber(this.rating)) {
-            this._sendFeedback(this.$('textarea').val());
-        }
-    },
-    /**
-     * @private
-     * @param {MouseEvent} ev
-     */
-    _onClickSmiley: function (ev) {
-        this.rating = parseInt($(ev.currentTarget).data('value'));
-        this.$('.o_livechat_rating_choices img').removeClass('selected');
-        this.$('.o_livechat_rating_choices img[data-value="' + this.rating + '"]').addClass('selected');
-
-        // only display textearea if bad smiley selected
-        if (this.rating !== 5) {
-            this._sendFeedback();
-            this.$('.o_livechat_rating_reason').show();
-        } else {
-            this.$('.o_livechat_rating_reason').hide();
-            this._showThanksMessage();
-            this._sendFeedback();
-        }
-    },
-    /**
-    * @private
-    */
-    _onEmailChat: function () {
-        var self = this;
-        var $email = this.$('#o_email');
-
-        if (utils.is_email($email.val())) {
-            $email.removeAttr('title').removeClass('is-invalid').prop('disabled', true);
-            this.$('.o_email_chat_button').prop('disabled', true);
-            this._rpc({
-                route: '/im_livechat/email_livechat_transcript',
-                params: {
-                    uuid: this._livechat.getUUID(),
-                    email: $email.val(),
-                }
-            }).then(function () {
-                self.$('.o_livechat_email').html($('<strong />', { text: _t('Conversation Sent') }));
-            }).guardedCatch(function () {
-                self.$('.o_livechat_email').hide();
-                self.$('.o_livechat_email_error').show();
-            });
-        } else {
-            $email.addClass('is-invalid').prop('title', _t('Invalid email address'));
-        }
-    },
-    /**
-    * @private
-    */
-    _onTryAgain: function () {
-        this.$('#o_email').prop('disabled', false);
-        this.$('.o_email_chat_button').prop('disabled', false);
-        this.$('.o_livechat_email_error').hide();
-        this.$('.o_livechat_email').show();
-    },
-});
-
-return {
-    LivechatButton: LivechatButton,
-    Feedback: Feedback,
-};
+return LivechatButton;
 
 });
-
